Refactor to satisfy the linter overlords
